requirements.txt
streamlit
langchain
langchain-core
langchain-text-splitters
langchain-huggingface
sentence-transformers
torchvision
langchain-google-genai
langchain-pinecone
pinecone
pypdf
python-dotenv
requests
